GoogleAut component created, the useContext logic in LoginForm was moved to this new component
