+++ src/components/GoogleAuth/GoogleAuth.module.css
@@ -1,0 +1,5 @@
+.googleButtonContainer{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
