+++ src/components/LoginForm/LoginForm.module.css
@@ -71,9 +71,3 @@
         margin-bottom: 15px;
     }
 }
-
-.googleButtonContainer{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
