+++ src/context/ProtectedWithGoogleAuth.js
@@ -2,9 +2,9 @@ import { Navigate } from "react-router-dom";
 import { useGoogleAuth } from "./GoogleAuthContext";
 
 const Protected = ({ children }) => {
-    const { user } = useGoogleAuth();
+    const { googleUser } = useGoogleAuth();
 
-    if (!user) return <Navigate to="/login" />; // If user is not logged in, redirect them to the login page.
+    if (!googleUser) return <Navigate to="/login" />; // If user is not logged in, redirect them to the login page.
 
     return children; // Otherwise, show the protected content
 };
